+++ tests/test_elements/test_ui_text_box.py
@@ -6,6 +6,7 @@ import pytest
 import i18n
 
 import pygame_gui
+from pygame_gui.core.utility import clipboard_paste
 from pygame_gui.elements.ui_text_box import UITextBox
 from pygame_gui.ui_manager import UIManager
 from tests.shared_comparators import compare_surfaces
@@ -968,6 +969,26 @@ class TestUITextBox:
                              manager=default_ui_manager)
         assert text_box.image is not None
         assert len(text_box.text_box_layout.layout_rows) == 3
+
+    def test_process_event_text_ctrl_c(self, _init_pygame: None,
+                                       _display_surface_return_none: None):
+        manager = UIManager((800, 600), os.path.join("tests", "data",
+                                                     "themes",
+                                                     "ui_text_entry_line_non_default_2.json"))
+        text_entry = UITextBox(html_text="dan",
+                               relative_rect=pygame.Rect(100, 100, 200, 30),
+                               manager=manager)
+
+        text_entry.focus()
+        text_entry.select_range = [0, 3]
+
+        processed_key_event = text_entry.process_event(pygame.event.Event(pygame.KEYDOWN,
+                                                                          {'key': pygame.K_c,
+                                                                           'mod': pygame.KMOD_CTRL,
+                                                                           'unicode': 'c'}))
+        text_entry.cursor_on = True
+
+        assert processed_key_event and clipboard_paste() == 'dan'
 
 
 if __name__ == '__main__':
